+++ lib/heuristics/scheduling_heuristic.rb
@@ -1209,6 +1209,7 @@ module Wrappers
 
       # configuration
       route_vrp.schedule_range_indices = nil
+      route_vrp.schedule_start_date = nil
 
       route_vrp.resolution_minimum_duration = 100
       route_vrp.resolution_time_out_multiplier = 5